add support for 64-bit ARM
Add a third processor type to the .so, and fix first compile-time
warnings and then a few dumb bugs based on assumptions about ptr sizes.
Works to play networked games and browse wordlists, but is not
extensively tested and needs to be before release.

+++ xwords4/android/app/src/main/java/org/eehouse/android/xw4/DictBrowseDelegate.java
@@ -54,7 +54,7 @@ public class DictBrowseDelegate extends DelegateBase
     private static final int MIN_LEN = 2;
 
     private Activity m_activity;
-    private int m_dictClosure = 0;
+    private long m_dictClosure = 0L;
     private int m_lang;
     private String m_name;
     private DictUtils.DictLoc m_loc;
